+++ components/elements/cards/ShareProgressCard/index.tsx
@@ -1,0 +1,12 @@
+import { SpeakerphoneIcon } from '@heroicons/react/outline';
+
+export const ShareProgressCard = () => {
+  return (
+    <section>
+      <div className='flex items-center'>
+        <SpeakerphoneIcon className='h-8' />
+        <h3>Share your progress</h3>
+      </div>
+    </section>
+  );
+};
