+++ components/elements/index.tsx
@@ -1,8 +1,10 @@
-import { ThemeSwitcherButton } from './ThemeSwitcherButton';
-
-export { ProjectCard, ArticleCard, ShareProgressCard } from './Card';
+export {
+  ArticleCard,
+  ProjectCard,
+  ProgressCard,
+  ShareProgressCard,
+} from './cards';
 export { Layout } from './Layout';
 export { Profile } from './Profile';
-export { Progress } from './Progress';
 export { SearchBar } from './SearchBar';
 export { ThemeSwitcherButton } from './ThemeSwitcherButton';
