+++ components/sections/ProgressGroup/index.tsx
@@ -1,9 +1,9 @@
-import { Progress } from '@components/elements';
+import { ProgressCard } from '@components/elements';
 
 export const ProgressGroup = () => {
   return (
     <ul className='grid w-full gap-12 mx-auto lg:mx-0 lg:max-w-full lg:grid-cols-2 lg:grid-rows-2'>
-      <Progress
+      <ProgressCard
         name='Linux & networking'
         courseID='sand8cs9a8'
         instructor='Bill Nye'
@@ -11,7 +11,7 @@ export const ProgressGroup = () => {
         progress={0.5}
         url='intro-to-networking'
       />
-      <Progress
+      <ProgressCard
         name='Programming & Scripting'
         courseID='caiushcsa8231'
         instructor='Michael Myers'
@@ -19,7 +19,7 @@ export const ProgressGroup = () => {
         progress={0.33}
         url='automation-with-python'
       />
-      <Progress
+      <ProgressCard
         name='Cloud Platforms'
         courseID='uhsaidoua823'
         instructor='Corey Quinn'
@@ -27,7 +27,7 @@ export const ProgressGroup = () => {
         progress={0.0}
         url='cloud-service-providers'
       />
-      <Progress
+      <ProgressCard
         name='DevOps Fundamentals'
         courseID='nidas9213'
         instructor='Paris Hilton'
